refactor: use BIGINT/i64 for users id and AppClaims.sub
- Migration: SERIAL -> BIGINT GENERATED ALWAYS AS IDENTITY
- User model: id i32 -> i64
- AppClaims: sub i32 -> i64

+++ src/core/middleware/auth_middleware.zig
@@ -8,7 +8,7 @@ const NextFn = spider.web.NextFn;
 const auth = spider.auth;
 
 pub const AppClaims = struct {
-    sub: i32,
+    sub: i64,
     email: []const u8,
     name: []const u8,
     locale: []const u8,
